Fix information. But.. I have to check no fix file. becuase when i update from MAC. I found un-update file..(or... re-file..)

+++ src/main/java/com/hmh/mmp/dto/bank/BankSaveDTO.java
@@ -14,14 +14,15 @@ import javax.validation.constraints.NotBlank;
 public class BankSaveDTO {
 
     private Long memberId;
+
     @Column(length = 50)
     @NotBlank(message = "계좌명은 필수입니다.") //
     private String bankName; // 계좌 명
     private String bankAccount; // 계좌번호
     private String bankMemo; // 메모
 
-    private long totalAsset; // 총 잔액
-    private Double bankRate; // 통장 이자율
+    private double totalAsset; // 총 잔액
+    private double bankRate; // 통장 이자율
     private MultipartFile bankPhoto; // 사진 항목
     private String bankPhotoName;
